+++ app/hashing/items.json
@@ -1,0 +1,7 @@
+[
+  {
+    "id": 1,
+    "label": "String Hashing",
+    "href": "/hashing"
+  }
+]
