+++ app/hashing/page.tsx
@@ -29,24 +29,22 @@ export default function Home() {
       <Head>
         <title>Digestify - Hashing Tools</title>
       </Head>
-      <main className="flex-1 flex flex-col items-center">
-        <section className="flex flex-col max-w-5xl w-full gap-10 justify-center my-0 px-12">
-          <span className="flex items-center gap-2 text-3xl mt-12 text-center justify-center">
-            Hash your data at lightspeed!
-          </span>
-          <div className="grid grid-cols-1 md:grid-cols-2 gap-4 items-center w-full">
-            <InputWrapper
-              value={inputValue}
-              onChange={(value) => updateValue(value)}
-              className="flex-1 min-h-52 h-auto ring-1 ring-border rounded-md resize-none focus:outline-none focus-visible:ring-2 focus-visible:ring-primary hover:ring-primary transition cursor-default"
-              placeholder="Paste a hashed string"
-            />
-            <OutputWrapper className="w-full min-h-52 h-auto bg-white dark:bg-[#121212] ring-1 ring-border rounded-md hover:ring-primary transition cursor-default">
-              {hashedValue}
-            </OutputWrapper>
-          </div>
-        </section>
-      </main>
+      <section className="flex flex-col max-w-5xl w-full gap-10 justify-center py-16 px-12">
+        <span className="flex items-center gap-2 text-3xl mt-12 text-center justify-center">
+          Hash your data at lightspeed!
+        </span>
+        <div className="grid grid-cols-1 md:grid-cols-2 gap-4 items-center w-full">
+          <InputWrapper
+            value={inputValue}
+            onChange={(value) => updateValue(value)}
+            className="flex-1 min-h-52 h-auto ring-1 ring-border rounded-md resize-none focus:outline-none focus-visible:ring-2 focus-visible:ring-primary hover:ring-primary transition cursor-default"
+            placeholder="Paste a hashed string"
+          />
+          <OutputWrapper className="w-full min-h-52 h-auto bg-white dark:bg-[#121212] ring-1 ring-border rounded-md hover:ring-primary transition cursor-default">
+            {hashedValue}
+          </OutputWrapper>
+        </div>
+      </section>
     </>
   )
 }
