--- types/base32-js.d.ts
@@ -1,1 +1,0 @@
-declare module 'base32.js'
